Refactor name detection screen
