+++ views/detection_config_screen.py
