+++ views/detection_config_screen.ui
